OPENGL: Add some support for multiple shaders
A "default" shader is used to scale the overlay and cursor. The special
shader is used only on the game screen. It looks for all .shader files
and tries to parse and compile them.

Currently there is no way to switch shaders and the "default" shader is
the same as the special one.